+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/instructions/ArrayAccessInstruction.java
@@ -4,7 +4,9 @@ import info.teksol.mc.mindcode.logic.arguments.LogicArray;
 import info.teksol.mc.mindcode.logic.arguments.LogicValue;
 import info.teksol.mc.mindcode.logic.arguments.arrays.ArrayConstructor;
 import info.teksol.mc.mindcode.logic.arguments.arrays.ArrayConstructor.AccessType;
+import org.jspecify.annotations.NullMarked;
 
+@NullMarked
 public interface ArrayAccessInstruction extends LogicInstruction {
 
     default LogicArray getArray() {