+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/instructions/ArrayOrganization.java
@@ -1,11 +1,14 @@
 package info.teksol.mc.mindcode.logic.instructions;
 
 import info.teksol.mc.mindcode.logic.arguments.arrays.*;
+import org.jspecify.annotations.NullMarked;
+import org.jspecify.annotations.Nullable;
 
 import java.util.function.Function;
 
+@NullMarked
 public enum ArrayOrganization {
-    NONE                ("none",        _ -> null),
+    NONE                ("none",        null),
     INTERNAL_REGULAR    ("regular",     RegularArrayConstructor::new),
     INTERNAL_SIZE1      ("short:1",     ArraySize1Constructor::new),
     INTERNAL_SIZE2      ("short:2",     ArraySize2Constructor::new),
@@ -14,9 +17,9 @@ public enum ArrayOrganization {
     ;
 
     private final String name;
-    private final Function<ArrayAccessInstruction, ArrayConstructor> expander;
+    private final @Nullable Function<ArrayAccessInstruction, ArrayConstructor> expander;
 
-    ArrayOrganization(String name, Function<ArrayAccessInstruction, ArrayConstructor> expander) {
+    ArrayOrganization(String name, @Nullable Function<ArrayAccessInstruction, ArrayConstructor> expander) {
         this.name = name;
         this.expander = expander;
     }
@@ -26,6 +29,7 @@ public enum ArrayOrganization {
     }
 
     public ArrayConstructor getExpander(ArrayAccessInstruction instruction) {
+        assert expander != null;
         return expander.apply(instruction);
     }
 }